social-to-lead-agent/requirements.txt
langchain>=0.1.0
langgraph>=0.0.40
langchain-google-genai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.9.0
streamlit>=1.35.0
